Add contactPolicy for methods

+++ app/Http/Controllers/ContactController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Resources\ContactResource;
 use App\Models\Contact;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
@@ -13,19 +14,27 @@ class ContactController extends Controller
      */
     public function index(Request $request)
     {
+        if ($request->user()->cannot('viewAny', Contact::class)) {
+            abort(403);
+        }
+
         $contacts = Contact::query();
         if ($request->user()->role == 'user') {
             $contacts = $contacts->where('user_id', $request->user()->id);
         }
         $contacts = $contacts->get();
-        return Inertia::render('contacts/index', ['contacts' => $contacts]);
+        return Inertia::render('contacts/index', ['contacts' => ContactResource::collection($contacts)]);
     }
 
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
+    public function create(Request $request)
     {
+        if ($request->user()->cannot('create', Contact::class)) {
+            abort(403);
+        }
+
         return Inertia::render('contacts/create');
     }
 
@@ -34,6 +43,10 @@ class ContactController extends Controller
      */
     public function store(Request $request)
     {
+        if ($request->user()->cannot('viewAny', Contact::class)) {
+            abort(403);
+        }
+
         $validatedData = $request->validate([
             'name' => 'required',
             'email' => 'nullable|email|unique:contacts,email',
@@ -57,8 +70,12 @@ class ContactController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Contact $contact)
+    public function edit(Request $request, Contact $contact)
     {
+        if ($request->user()->cannot('update', $contact)) {
+            abort(403);
+        }
+
         return Inertia::render('contacts/edit', [
             'contact' => $contact
         ]);
@@ -69,21 +86,30 @@ class ContactController extends Controller
      */
     public function update(Request $request, Contact $contact)
     {
+        if ($request->user()->cannot('update', $contact)) {
+            abort(403);
+        }
+
         $validatedData = $request->validate([
             'name' => 'required',
             'email' => 'nullable|email|unique:contacts,email,' . $contact->id,
             'mobile' => 'required|numeric|digits:9|unique:contacts,mobile,' . $contact->id,
             'address' => 'nullable'
         ]);
-        Contact::updated($validatedData);
+        $contact->update($validatedData);
         return redirect()->route('contacts.index')->with('success', 'Contact updated successfully');
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Contact $contact)
+    public function destroy(Request $request, Contact $contact)
     {
+
+        if ($request->user()->cannot('delete', $contact)) {
+            abort(403);
+        }
+
         $contact->delete();
         return redirect()->route('contacts.index')->with('success', 'Contact deleted successfully');
     }
